Consolidated admin components into directory, updated imports
