+++ client/components/admin/AdminSingleProduct.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const AdminSingleProduct = props => {
+  return (
+    <div>
+      <div>This will be something soon</div>
+      <div>{`Param: ${props.match.params.productId}`}</div>
+    </div>
+  )
+}
+
+export default AdminSingleProduct
