+++ client/components/admin/index.js
@@ -1,0 +1,7 @@
+export {default as AdminDashboard} from './AdminDashboard'
+export {default as AdminDashboardRoutes} from './AdminDashboardRoutes'
+export {default as AdminProductsList} from './AdminProductsList'
+export {default as AdminProductsListRow} from './AdminProductsListRow'
+export {default as AdminSingleProduct} from './AdminSingleProduct'
+export {default as AdminUsersList} from './AdminUsersList'
+export {default as AdminUserTableRow} from './AdminUserTableRow'
